// src/helper/types/lido_terra_hub/unbond_requests_response.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type HumanAddr = string;
export type Uint128 = string;

export interface UnbondRequestsResponse {
  address: HumanAddr;
  requests: [number, Uint128, Uint128][];
  [k: string]: unknown;
}